state pattern updated
