--- src/main/java/com/practice/designpattern/behavioral/memento/EditorState.java
@@ -1,4 +1,0 @@
-package com.practice.designpattern.behavioral.memento;
-
-public record EditorState(String content) {
-}
