+++ src/main/java/com/practice/designpattern/behavioral/memento/good/EditorState.java
@@ -1,0 +1,4 @@
+package com.practice.designpattern.behavioral.memento.good;
+
+public record EditorState(String content) {
+}
