+++ src/main/java/com/practice/designpattern/behavioral/state/good/Canvas.java
@@ -1,11 +1,13 @@
-package com.practice.designpattern.behavioral.state;
+package com.practice.designpattern.behavioral.state.good;
 
+import com.practice.designpattern.behavioral.state.good.tools.Tool;
 import lombok.Getter;
 import lombok.Setter;
 
 public class Canvas {
 
-    @Getter @Setter
+    @Getter
+    @Setter
     private Tool currentTool;
 
     public void mouseDown() {
