+++ src/main/java/com/practice/designpattern/behavioral/state/good/tools/Brush.java
@@ -1,0 +1,15 @@
+package com.practice.designpattern.behavioral.state.good.tools;
+
+public class Brush implements Tool {
+
+    @Override
+    public void mouseDown() {
+        System.out.println("Brush Icon");
+    }
+
+    @Override
+    public void mouseUp() {
+        System.out.println("Draw a line");
+    }
+
+}
