+++ src/main/java/com/practice/designpattern/behavioral/state/good/tools/Selection.java
@@ -1,0 +1,15 @@
+package com.practice.designpattern.behavioral.state.good.tools;
+
+public class Selection implements Tool {
+
+    @Override
+    public void mouseDown() {
+        System.out.println("Selection Icon");
+    }
+
+    @Override
+    public void mouseUp() {
+        System.out.println("Draw dashed rectangle");
+    }
+
+}
