+++ src/main/java/com/practice/designpattern/behavioral/state/good/tools/Tool.java
@@ -1,0 +1,9 @@
+package com.practice.designpattern.behavioral.state.good.tools;
+
+public interface Tool {
+
+    void mouseDown();
+
+    void mouseUp();
+}
+
