+++ src/main/java/com/practice/educative/grokking/binarytree/FindAllTreePaths.java
@@ -4,5 +4,6 @@ import java.util.List;
 
 public interface FindAllTreePaths {
 
-    public abstract List<List<Integer>> findPaths(TreeNode<Integer> root, Integer sum);
+    List<List<Integer>> findPaths(TreeNode<Integer> root, Integer sum);
+
 }
