+++ src/test/java/com/practice/java/stream/exercise/StreamApiTest.java
@@ -3,25 +3,20 @@
  */
 package com.practice.java.stream.exercise;
 
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.log;
+import com.practice.data.model.Customer;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
-import java.util.Random;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import com.practice.data.model.Customer;
-
-import lombok.extern.slf4j.Slf4j;
 
 /**
  * @author swjana
